Moved Scheduler back to usecases
