+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/AutoAgenda.iml" filepath="$PROJECT_DIR$/AutoAgenda.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/autoagenda.iml" filepath="$PROJECT_DIR$/autoagenda.iml" />
       <module fileurl="file://$PROJECT_DIR$/autoagendacore/autoagendacore.iml" filepath="$PROJECT_DIR$/autoagendacore/autoagendacore.iml" />
     </modules>
   </component>
